Follow refactor

+++ routes/request.js
@@ -342,9 +342,11 @@ async function sendRequest(authorId, foreignId, res) {
 async function apideleteRequest(authorId, foreignId, res) {
     const actor = await Author.findOne({_id: authorId});  
     const object = await Author.findOne({_id: foreignId});
+    if (!actor && !object) { return 500 }
     let summary = '';
 
     const request = await Request.findOneAndDelete({senderId: actor._id, receiverId: object._id}); 
+    if (!request) { return 500 }
     summary = actor.username + " wants to undo " + request.type + " request to " + object.username;  
 
     return {
@@ -355,6 +357,22 @@ async function apideleteRequest(authorId, foreignId, res) {
     }
 }
 
+async function getRequests(authorId) {
+    await Request.find({object: authorId}, function(err, requests){
+        console.log("Debug: Requests exists");
+        return requests
+    }).clone()
+}
+
+async function getRequest(authorId, foreignId) {
+    await Request.findOne({actor: authorId, object: foreignId}, function(err, request){
+        if (!request) { console.log('Debug: Does not exist') }
+        return res.json({
+            request: request
+        })
+    }).clone()
+}
+
 module.exports={
     saveRequest,
     deleteRequest,
@@ -362,5 +380,7 @@ module.exports={
     findAllRequests,
     senderAdded,
     sendRequest,
-    apideleteRequest
+    apideleteRequest,
+    getRequests,
+    getRequest
 }
